requirements.txt
aerich==0.9.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.1
asyncclick==8.3.0.7
asyncpg==0.30.0
attrs==25.4.0
colorama==0.4.6
dictdiffer==0.9.0
frozenlist==1.8.0
idna==3.11
iso8601==2.1.0
magic-filter==1.0.12
maxapi==0.9.7
multidict==6.7.0
propcache==0.4.1
puremagic==1.30
pydantic==2.12.4
pydantic_core==2.41.5
pypika-tortoise==0.6.2
python-dotenv==1.2.1
pytz==2025.2
sniffio==1.3.1
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
yarl==1.22.0
